+++ src/LiveSplit.Racetime/Controller/IAuthentificationSettings.cs
