+++ src/LiveSplit.Racetime/Controller/RTAuthentificationSettings.cs
