+++ src/LiveSplit.Racetime/Controller/RacetimeAuthenticator.cs
