+++ src/LiveSplit.Racetime/Model/ChatMessage.cs
