+++ src/LiveSplit.Racetime/Model/MessageType.cs
