+++ src/LiveSplit.Racetime/Model/RTModelBase.cs
