+++ src/LiveSplit.Racetime/Model/Race.cs
