+++ src/LiveSplit.Racetime/Model/RaceState.cs
