+++ src/LiveSplit.Racetime/Model/RacetimeUser.cs
