+++ src/LiveSplit.Racetime/Model/SplitUpdate.cs
