+++ src/LiveSplit.Racetime/Model/UserRole.cs
