+++ src/LiveSplit.Racetime/Model/UserStatus.cs
