+++ src/LiveSplit.Racetime/Properties/Resources.resx
