+++ src/LiveSplit.Racetime/RacetimeAPI.cs
