+++ src/LiveSplit.Racetime/RacetimeComparisonGenerator.cs
