+++ src/LiveSplit.Racetime/RacetimeSettings.cs
