+++ src/LiveSplit.Racetime/RacetimeSettingsControl.cs
