+++ src/LiveSplit.Racetime/RacetimeSettingsControl.resx
