+++ src/LiveSplit.Racetime/View/ChannelForm.cs
