+++ src/LiveSplit.Racetime/View/ChannelForm.resx
